fix : compile warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@28910 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/bar/barBkupLogger.cpp
@@ -1430,7 +1430,7 @@ namespace engine
             else
             {
                onceLen = _curOffset + DMS_SEGMENT_SZ ;
-               if ( onceLen > pLobData->getFileSz() )
+               if ( onceLen > (UINT64)pLobData->getFileSz() )
                {
                   onceLen = pLobData->getFileSz() ;
                }
